More simplifies.

+++ configure.d/inferno.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
-
-rm -f "$INFERNO/usr/$USER"
-ln -s "$HOME" "$INFERNO/usr/$USER"
+if test -d "$INFERNO" ; then
+	mkdir -p "$INFERNO/usr/"
+	rm -f "$INFERNO/usr/$USER"
+	ln -s "$HOME" "$INFERNO/usr/$USER"
+fi
